+++ src/sx126x_driver_version.h
@@ -55,8 +55,8 @@ extern "C" {
  */
 
 #define SX126X_DRIVER_VERSION_MAJOR 2
-#define SX126X_DRIVER_VERSION_MINOR 2
-#define SX126X_DRIVER_VERSION_PATCH 0
+#define SX126X_DRIVER_VERSION_MINOR 3
+#define SX126X_DRIVER_VERSION_PATCH 2
 
 /*
  * -----------------------------------------------------------------------------